feat!: fetch weather data with test lat long

+++ app/src/main/java/com/example/superweather/data/model/weather/Main.kt
@@ -1,7 +1,7 @@
 package com.example.superweather.data.model.weather
 
 data class Main(
-    val feels_like: Int,
+    val feels_like: Double,
     val humidity: Int,
     val pressure: Int,
     val temp: Double,
